Closer to spec, with tests
On a separate branch, I managed to get the test cases given by the HAML spec
working, but destabilized another test case (added here as
`test_inner_and_outer_whitespace_advanced`) which is one of the core reasons I
started fooling around with this library in the first place. For now, I've got
what I need and the rest of the library is still functioning, so I'm content
with "close enough".